Move to UserUtil for profile configuration level check
